[FIX/#28] AddressCard 컴포넌트 분리
