+++ app/src/main/java/com/sopt/shinmungo/presentation/map/component/AddressCard.kt
@@ -1,0 +1,62 @@
+package com.sopt.shinmungo.presentation.map.component
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.unit.dp
+import com.sopt.shinmungo.R
+import com.sopt.shinmungo.core.designsystem.theme.ShinMunGoTheme
+
+
+@Composable
+fun AddressCard(
+    modifier: Modifier = Modifier,
+    address: String,
+) {
+    Card(
+        modifier = modifier.fillMaxWidth(),
+        shape = RoundedCornerShape(5.dp),
+        colors = CardDefaults.cardColors(
+            containerColor = Color.White
+        ),
+        border = BorderStroke(
+            width = 1.dp,
+            color = ShinMunGoTheme.color.gray5
+        )
+    ) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(16.dp),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.spacedBy(8.dp)
+        ) {
+            Icon(
+                imageVector = ImageVector.vectorResource(id = R.drawable.ic_location_line_black_16px),
+                contentDescription = "위치",
+                tint = ShinMunGoTheme.color.primary,
+                modifier = Modifier.size(16.dp)
+            )
+
+            Text(
+                text = address,
+                style = ShinMunGoTheme.typography.body8,
+                color = ShinMunGoTheme.color.primary
+            )
+        }
+    }
+}
